Moved ITradeManager into Interfaces
