+++ src/Mynt.AspNetCore.Host/Hosting/MyntHostedService.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Mynt.Core.TradeManagers;
+using Mynt.Core.Interfaces;
 
 namespace Mynt.AspNetCore.Host.Hosting
 {
